Refactor App.tsx into modular components with improved structure
