+++ workspace/react-app/src/constants/colors.ts
@@ -1,0 +1,5 @@
+export const COLORS = [
+  '#ffffff', '#616161', '#f0f0f0', '#5b5b5b', '#222222', '#fc8d74', '#432d26', '#eead91', '#806355', '#382d21',
+  '#faef93', '#aeba5e', '#8aa140', '#1f6522', '#13afa2', '#b8d5d7', '#15aeda', '#a5def8', '#0f77c0', '#3469b7',
+  '#c50404'
+]
